Split source control tag hit testing
